Replace uint64 with int64 for artist_id

+++ internal/fetcher/services/itunes/itunes.go
@@ -72,7 +72,7 @@ func (f *Fetcher) fetchWorker(id int, artists <-chan *artists.StoreInfo, wg *syn
 			err = db.DbMgr.EnsureReleaseExists(&db.Release{
 				StoreName: f.GetStoreName(),
 				StoreID:   release.ID,
-				ArtistID:  uint64(artist.ArtistID),
+				ArtistID:  artist.ArtistID,
 				Title:     removeAlbumType(release.Attributes.Name),
 				Poster:    release.Attributes.Artwork.GetLink(posterWidth, posterHeight),
 				Released:  release.Attributes.ReleaseDate.Value,
